// prodis-prodiswcl/src/app/modules/prodisapi/model/provincia.ts
/*
* SPDX-FileCopyrightText: Copyright 2022 | Regione Piemonte
* SPDX-License-Identifier: EUPL-1.2
*/
/**
 * Prodis
 * API per il backend della suite di Prodis.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Comune } from './comune';
import { Regione } from './regione';


export interface Provincia { 
    codProvinciaMin?: string;
    comunes?: Array<Comune>;
    dsProTProvincia?: string;
    dsTarga?: string;
    dtFine?: Date;
    dtInizio?: Date;
    dtTmst?: Date;
    id?: number;
    regione?: Regione;
}